% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsearch.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get information about data sets in a package.}
\usage{
get_dataset_info(package_name = "datasets", types = c("factor", "numeric",
  "ordered", "integer", "ts"))
}
\arguments{
\item{package_name}{The name of an R package. The package must be loaded.}

\item{types}{A vector of variable types that should be counted.}
}
\value{
A data frame summarizing the data sets in a package
}
\description{
Get information about data sets in a package.
}
